+++ kindle/WebLaunch/settings.js
@@ -1,9 +1,0 @@
-var settings = {
-    url: 'file:///mnt/us/extensions/WebLaunch/bin/dash.html',
-    title: 'Ink Dashboard',
-    hideStatusbar: true,
-    enableWireless: false,
-    powerButtonClose: true,
-    enablePreventScreenSaver: true,
-    landscape: false
-};